+++ docs/pages/index.mdx
@@ -1,6 +1,5 @@
-import { useData } from 'nextra/data';
-import { getStargazersCount } from '../utils/stargazers';
 import { Homepage } from '../components/Homepage';
+import { getStargazersCount } from '../utils/getStargazersCount';
 
 export async function getStaticProps() {
   const { version } = require('../../lerna.json');
