+++ docs/utils/useStargazersCount.ts
@@ -1,17 +1,5 @@
 import React from 'react';
-
-const repoUrl = 'https://api.github.com/repos/react-cosmos/react-cosmos';
-const starsFallback = 7905;
-
-export async function getStargazersCount() {
-  try {
-    const res = await fetch(repoUrl);
-    const repo = await res.json();
-    return repo.stargazers_count ?? 0;
-  } catch (err) {
-    return starsFallback;
-  }
-}
+import { getStargazersCount } from './getStargazersCount';
 
 const animationDuration = 1000; // ms
